数据格式调整

+++ tmp/thrift/Cat.thrift
@@ -2,7 +2,7 @@
 service CatText{
 
 void  LoadResource(),
-i32 CatNumber(1:string doc),
+i64 CatNumber(1:string doc),
 string CatName(1:string doc),
 string TextFeature(1:string doc),
 }
